+++ test/compojure/api/coercion_test.clj
@@ -90,7 +90,7 @@
         (fact "by default, applies body coercion (to set)"
           (let [app (api
                       beer-route)]
-            (let [[status body] (post* app "/beer" (json {:beers ["ipa" "apa" "ipa"]}))]
+            (let [[status body] (post* app "/beer" (json-string {:beers ["ipa" "apa" "ipa"]}))]
               status => 200
               body => {:beers ["ipa" "apa"]})))
 
@@ -99,13 +99,13 @@
                 app (api
                       {:coercion no-body-coercion}
                       beer-route)]
-            (let [[status body] (post* app "/beer" (json {:beers ["ipa" "apa" "ipa"]}))]
+            (let [[status body] (post* app "/beer" (json-string {:beers ["ipa" "apa" "ipa"]}))]
               status => 200
               body => {:beers ["ipa" "apa" "ipa"]}))
           (let [app (api
                       {:coercion nil}
                       beer-route)]
-            (let [[status body] (post* app "/beer" (json {:beers ["ipa" "apa" "ipa"]}))]
+            (let [[status body] (post* app "/beer" (json-string {:beers ["ipa" "apa" "ipa"]}))]
               status => 200
               body => {:beers ["ipa" "apa" "ipa"]})))
 
@@ -114,7 +114,7 @@
                 app (api
                       {:coercion nop-body-coercion}
                       beer-route)]
-            (post* app "/beer" (json {:beers ["ipa" "apa" "ipa"]})) => (fails-with 400)))))
+            (post* app "/beer" (json-string {:beers ["ipa" "apa" "ipa"]})) => (fails-with 400)))))
 
     (fact "query coercion"
       (let [query-route (GET "/query" []